Updating DML messages to use DMLClientAPI. Also required some changes to related classes.
The DMLClientAPI scan methods require a Session, so I've snuck that into the Messages (in DMLRequest) for now, to stay out of Amos' changes. Similarly, cursor lifecycle is being handled for now by static methods in ScanRowsRequest.java. Both of these are temporary, designed to get the code working, be easy to replace, and hopefully avoid code conflicts with Amos' network/session code.