Making the HapiProcessor's output more flexible. Instead of HapiProcessor.processRequest returning a String, it takes as an argument a (new) interface HapiProcessor.Outputter<T>, and returns a T. This lets you return a String, byte[], whatever.
I'm including a few outputters to start with. MemcacheServiceImpl maintains an atomic enum, where each element corresponds to an Outputter<byte[]>, and JMX options to query available enums and set the active one. This is then passed to AkibanCommandHandler. End result is that by setting the enum you can dynamically set the output format.