Add map of default type parameter values to Types class next to static list of types:     - Moves DECIMAL specifics out of Column methods, slightly more general and now only in a single place     - Removed deprecated and unused Column constructors while I was there     - Supports default values for types that have them (char, decimal, binary, and bit).
Add special parse time conversions for BIT(0) and FLOAT(P) to schemaDef.addColumn().

Change Column.paramCheck range check to <0 instead of <1 as 0 is valid for a number of types. This eliminated the need for one of the tests in AtomicSchemaChanges.

Add tests for new functionality and simplify/shorten others where possible in CreateTableIT.