Fix parsing of DDL statements that contain comments:   - Remove group comment cruft from grammar and SchemaDef   - Direct all comments in grammar to hidden channel   - Fix single line comment lexar rule to handle all line endings   - Rename comment lexer rule names so COMMENT token could be renamed COMMENT.
Ease back transformation that SchemaDef.canonicalStatement() does as it could actually take a valid DDL string and turn it into an invalid one (embedded -- comment, for example). Also change PSSMT to use SchemaDef.canonicalStatement() instead of DDLSource to a) use this improvement and b) remove another usage of deprecated DDLSource.