Fix the merge problems with trunk.
Remove the MemoryStore as a service. This was created to manage the MemoryTableFactories, but the design was changed to allow the UserTable (in the AIS) to hold these, and make the SchemaManager mange them. 

This causes the method to find the MemoryTableFactory for specific tables to change, which causes a cascade of minor refactoring changes.