Add TablesDDL support for creating and dropping (but not altering) a table with GENERATE IDENTITY.
The command has the TablesDDL create a "hidden" sequence for generating the values, which is then attached to the identity column of the table. Dropping the table also drops the sequence. 

Add tests for this at several points, including the AISMerge, AISBuilder and the TablesDDLIT to verify functionality. There probably could be more.