Updating PersistitKeyPValueSource to get the collated CString (in String form) and put it into the PValueSource's cache -- instead of getting that string and putting it in the PValueSource's raw value. This means you can only retrieve it by calling getObject(), instead of the usual getString(). Fix the few places in which this is correct behavior; in other places, trying to call getString() will correctly throw an exception.
This is a way of making explicit, in the runtime at least, the fact that the value of a collated string (e.g. an indexed VARCHAR) can't be treated as a normal String. If the collatation isRecoverable, we should be able to move the String back to the PValueSource's raw value. I'm not sure if we ever try to do this, but if we do, I'll find that and fix it in a separate commit.