Syncing the error cases for the two SpaceModificationHandlers.
This is done via two mechanisms: Firstly, one error check has been moved to SpaceDiff. Secondly, the errors reported from the SMHs are now all in a base class. So basically, as long as every SMH extends that base class (they both do right now) and never calls this.error(String), the errors should be in sync. This is the case as of this commit. To help remind developers to keep this system in place, the ASMH methods are all final.