Several tweaks to calls for open/close and adding more checks for cursor state.
Make sure the several places that implement a real version of jump set
the cursor state back to ACTIVE to allow correct reading of next().

catch two cases of close() which should be setIdle(). 

Group Lookup Default is still in a debugging state, but has several
important fixes to be included.