Update the PersistitGroupCursor to use the new expandRow() interfaces.
This in turn requires PersistitStore and PersistitStorageDescription to
implement the expandRow. 

This in turn requires the addition of RowValueEncoder() to encode/decode
the ValuesHolderRow to the PersistitValue. Added a test for this last
one. 

The PersistitStorageDescription, for the Row processing no longer
encodes the row as a RowData, instead the data is encoded/decoded as
separate values in the Value. No more RowData along that path. 

Testing still needs to occur to verify this is correct.