Use method of ValueNode
since VirtualColumnNode now overrides the base method, there is no
longer a need to cast here (yay).

That being said, the intention for this action is to avoid failing
CheckParserUsagesDT, which makes me a little uncomfortable,
and leaves a bunch of questions open, but as long as things work,
I feel ok. And RandomSemiJoinDT does run through a lot of subqueries.
Of course, none of those involve SELECT * which is when
VirtualColumnNode comes into play...